# internal/docs/swagger.yaml
definitions:
...
  dto.ConversationImportRequest:
    properties:
      content:
        type: string
      source:
        type: string
      tags:
        items:
          type: string
        type: array
    required:
    - content
    type: object
  dto.ConversationImportResponse:
    properties:
      import_id:
        type: integer
      message:
        type: string
      message_count:
        type: integer
    type: object
...
  dto.FileFragment:
    properties:
      file_name:
        type: string
      matched_terms:
        items:
          type: string
        type: array
      modified:
        type: string
      path:
        type: string
      preview:
        description: First 500 chars
        type: string
      size:
        type: integer
    type: object
  dto.FileScanRequest:
    properties:
      extensions:
        description: e.g., [".md", ".txt", ".go"]
        items:
          type: string
        type: array
      max_depth:
        description: Max directory depth
        type: integer
      max_results:
        description: Max files to return
        type: integer
      root_path:
        type: string
      search_terms:
        items:
          type: string
        type: array
    type: object
  dto.FileScanResponse:
    properties:
      files_found:
        items:
          $ref: '#/definitions/dto.FileFragment'
        type: array
      imported_count:
        type: integer
      scanned_paths:
        type: integer
      total_matches:
        type: integer
    type: object
  dto.ImportFragmentsRequest:
    properties:
      event_type:
        description: e.g., "solace_fragment"
        type: string
      file_paths:
        items:
          type: string
        type: array
    type: object
  dto.ImportFragmentsResponse:
    properties:
      errors:
        items:
          type: string
        type: array
      imported_count:
        type: integer
    type: object
  dto.LimitOrderRequest:
    properties:
      coin_id:
        type: string
      currency:
        type: string
      limit_price:
        type: number
      quantity:
        type: number
      side:
        type: string
      symbol:
        type: string
    type: object
...
  dto.MarketOrderRequest:
    properties:
      coin_id:
        type: string
      currency:
        type: string
      quantity:
        type: number
      side:
        type: string
      symbol:
        type: string
    required:
    - coin_id
    - currency
    - quantity
    - side
    - symbol
    type: object
  dto.MemoryLearnRequest:
    properties:
      event_type:
        type: string
      payload:
        additionalProperties: true
        type: object
      session_id:
        type: string
    required:
    - event_type
    - payload
    type: object
  dto.MemoryLearnResponse:
    properties:
      id:
        type: integer
      message:
        type: string
    type: object
  dto.MemoryRecallResponse:
    properties:
      event_type:
        type: string
      id:
        type: integer
      payload:
        additionalProperties: true
        type: object
      session_id:
        type: string
      timestamp:
        type: string
      user_id:
        type: integer
    type: object
...
  dto.TradeResponse:
    properties:
      coin_id:
        type: string
      created_at:
        type: string
      id:
        type: integer
      price:
        type: number
      quantity:
        type: number
      side:
        type: string
      status:
        type: string
      symbol:
        type: string
      type:
        type: string
      updated_at:
        type: string
      user_id:
        type: integer
    type: object
info:
  contact: {}
paths:
...
  /memory/import:
    post:
      consumes:
      - application/json
      description: Imports a pasted conversation or text into memory system
      parameters:
      - description: Conversation Import
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.ConversationImportRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.ConversationImportResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Import a conversation
      tags:
      - Memory
  /memory/learn:
    post:
      consumes:
      - application/json
      description: Stores a memory event for the user
      parameters:
      - description: Memory Event
        in: body
        name: memory
        required: true
        schema:
          $ref: '#/definitions/dto.MemoryLearnRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/dto.MemoryLearnResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Store a memory snapshot
      tags:
      - Memory
  /memory/recall:
    get:
      consumes:
      - application/json
      description: Retrieves memory snapshots for the user
      parameters:
      - default: 20
        description: Number of snapshots to retrieve
        in: query
        name: limit
        type: integer
      - description: Filter by event type
        in: query
        name: event_type
        type: string
      - description: Filter by session ID
        in: query
        name: session_id
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/dto.MemoryRecallResponse'
            type: array
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Recall memory snapshots
      tags:
      - Memory
  /scanner/import:
    post:
      consumes:
      - application/json
      description: Imports discovered fragments into memory_snapshots table
      parameters:
      - description: Import request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.ImportFragmentsRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.ImportFragmentsResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Import file fragments into memory
      tags:
      - Scanner
  /scanner/scan:
    post:
      consumes:
      - application/json
      description: Recursively scans filesystem for files containing search terms
      parameters:
      - description: Scan request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.FileScanRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.FileScanResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Scan filesystem for Solace/ARES fragments
      tags:
      - Scanner
  /scanner/solace:
    post:
      consumes:
      - application/json
      description: Scans C:\ProgramData\Solace\State and imports all files
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.ImportFragmentsResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Import Solace Delta 3.1 data
      tags:
      - Scanner
...
  /trades/history:
    get:
      parameters:
      - description: Number of trades
        in: query
        name: limit
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/dto.TradeResponse'
            type: array
      security:
      - BearerAuth: []
      summary: Get last N trades for user
      tags:
      - Trading
  /trades/limit:
    post:
      consumes:
      - application/json
      parameters:
      - description: Limit Order
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.LimitOrderRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.TradeResponse'
      security:
      - BearerAuth: []
      summary: Execute Limit Order
      tags:
      - Trading
  /trades/market:
    post:
      consumes:
      - application/json
      parameters:
      - description: Market Order
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.MarketOrderRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.TradeResponse'
      security:
      - BearerAuth: []
      summary: Execute Market Order
      tags:
      - Trading
  /trades/pending:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/dto.TradeResponse'
            type: array
      security:
      - BearerAuth: []
      summary: Get pending limit orders for user
      tags:
      - Trading
  /trades/performance:
    get:
      description: Calculate P&L and performance metrics (scaffold - uses hardcoded
        prices)
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Get trading performance stats
      tags:
      - Trading
...
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and your JWT token (e.g., "Bearer
      <token>")
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"